Added: modules utils.js and data.js in js folder
