+++ index.html
@@ -228,5 +228,7 @@
     </div>
   </template>
  <script src="/js/main.js"></script>
+ <script type="module" src="/js/data.js"></script>
+ <script type="module" src="/js/utils.js"></script>
 </body>
 </html>
