+++ js/data.js
@@ -1,0 +1,11 @@
+const TYPE = ['palace', 'flat' , 'house' , 'bungalow'];
+const CHECKIN = ['12:00', '13:00', '14:00'];
+const CHECKOUT = ['12:00', '13:00', '14:00'];
+const FEATURES = ['wifi', 'dishwasher', 'parking', 'washer', 'elevator', 'conditioner'];
+const PHOTOS = ['http://o0.github.io/assets/images/tokyo/hotel1.jpg', 'http://o0.github.io/assets/images/tokyo/hotel2.jpg', 'http://o0.github.io/assets/images/tokyo/hotel3.jpg'];
+
+
+
+const countOfArrays = 10;
+
+export {TYPE,CHECKIN,CHECKOUT,FEATURES,PHOTOS,countOfArrays};
